test(copilot): split provider test suite
