+++ Tests/OpenUsageTests/CopilotAuthStoreTests.swift
@@ -1,0 +1,129 @@
+import XCTest
+@testable import OpenUsage
+
+final class CopilotAuthStoreTests: XCTestCase {
+    func testReadsEditorAppsJSON() {
+        let store = CopilotAuthStore(
+            files: FakeFiles([
+                CopilotAuthStore.editorAppsPath: """
+                { "github.com:Iv1.abc123": { "user": "octocat", "oauth_token": "gho_editor" } }
+                """
+            ]),
+            keychain: FakeKeychain()
+        )
+
+        let token = store.loadToken()
+
+        XCTAssertEqual(token?.value, "gho_editor")
+    }
+
+    func testReadsGhHostsOAuthToken() {
+        let store = CopilotAuthStore(
+            files: FakeFiles([
+                CopilotAuthStore.ghHostsPath: """
+                github.com:
+                    git_protocol: https
+                    user: octocat
+                    oauth_token: gho_ghconfig
+                """
+            ]),
+            keychain: FakeKeychain()
+        )
+
+        let token = store.loadToken()
+
+        XCTAssertEqual(token?.value, "gho_ghconfig")
+    }
+
+    func testDecodesGoKeyringWrappedGhKeychainToken() {
+        let wrapped = "go-keyring-base64:" + Data("gho_keychain".utf8).base64EncodedString()
+        let store = CopilotAuthStore(files: FakeFiles(), keychain: FakeKeychain(wrapped))
+
+        let token = store.loadToken()
+
+        XCTAssertEqual(token?.value, "gho_keychain")
+    }
+
+    func testEditorConfigWinsOverKeychain() {
+        let store = CopilotAuthStore(
+            files: FakeFiles([
+                CopilotAuthStore.editorAppsPath: #"{ "github.com": { "oauth_token": "gho_editor" } }"#
+            ]),
+            keychain: FakeKeychain("go-keyring-base64:" + Data("gho_keychain".utf8).base64EncodedString())
+        )
+
+        // Editor config wins over the keychain: the editor token is returned, not the keychain one.
+        XCTAssertEqual(store.loadToken()?.value, "gho_editor")
+    }
+
+    func testReturnsNilWhenNoCredentials() {
+        let store = CopilotAuthStore(files: FakeFiles(), keychain: FakeKeychain())
+        XCTAssertNil(store.loadToken())
+    }
+
+    func testEditorConfigIgnoresNonGithubDotComHost() {
+        // An Enterprise-only editor config must not yield a token for api.github.com; the chain should
+        // fall through to the gh keychain (which here holds the real github.com token).
+        let store = CopilotAuthStore(
+            files: FakeFiles([
+                CopilotAuthStore.editorAppsPath: #"{ "ghe.corp.example:Iv1.x": { "oauth_token": "gho_enterprise" } }"#
+            ]),
+            keychain: FakeKeychain("go-keyring-base64:" + Data("gho_dotcom".utf8).base64EncodedString())
+        )
+
+        let token = store.loadToken()
+
+        XCTAssertEqual(token?.value, "gho_dotcom")
+    }
+
+    func testEditorConfigPicksGithubDotComAmongHosts() {
+        let store = CopilotAuthStore(
+            files: FakeFiles([
+                CopilotAuthStore.editorAppsPath: #"{ "ghe.corp.example:Iv1.x": { "oauth_token": "gho_ent" }, "github.com:Iv1.y": { "oauth_token": "gho_dotcom" } }"#
+            ]),
+            keychain: FakeKeychain()
+        )
+
+        XCTAssertEqual(store.loadToken()?.value, "gho_dotcom")
+    }
+
+    func testYamlValueIgnoresNestedUsersMap() {
+        let hosts = """
+        github.com:
+            users:
+                octocat:
+            user: octocat
+        """
+        XCTAssertEqual(CopilotAuthStore.yamlValue(hosts, key: "user"), "octocat")
+    }
+
+    func testYamlValueScopesToGithubDotComHost() {
+        // A GitHub Enterprise block precedes github.com; the github.com token must win.
+        let hosts = """
+        ghe.corp.example:
+            oauth_token: gho_enterprise
+            user: ent
+        github.com:
+            oauth_token: gho_dotcom
+            user: octocat
+        """
+        XCTAssertEqual(CopilotAuthStore.yamlValue(hosts, key: "oauth_token"), "gho_dotcom")
+        XCTAssertEqual(CopilotAuthStore.yamlValue(hosts, key: "user"), "octocat")
+    }
+
+    func testGhConfigPrefersGithubDotComTokenOverEnterprise() {
+        let store = CopilotAuthStore(
+            files: FakeFiles([
+                CopilotAuthStore.ghHostsPath: """
+                ghe.corp.example:
+                    oauth_token: gho_enterprise
+                github.com:
+                    oauth_token: gho_dotcom
+                """
+            ]),
+            keychain: FakeKeychain()
+        )
+
+        XCTAssertEqual(store.loadToken()?.value, "gho_dotcom")
+    }
+}
